DRC-2170: More cleaning and debugging and testing of QuadTree. Won't fixBamboo.

+++ SensorProcessing/src/us/ihmc/sensorProcessing/pointClouds/combinationQuadTreeOctTree/GroundOnlyQuadTree.java
@@ -459,7 +459,7 @@ public class GroundOnlyQuadTree extends HyperCubeTree<GroundAirDescriptor, Groun
       this.octree = octree;
    }
 
-   public void setHeightThreshold(float threshold)
+   public void setHeightThreshold(double threshold)
    {
       this.heightThreshold = threshold;
    }
